dedupe quotes
The poor man's check in shell:

    $ grep "\stext:" app/coins/btcQuotes.js \
      | sort | sed 's/  / /' \
      | uniq -c | grep -v "\s1"
          2                     text: "I'm sure that in 20 years there will either be very large transaction volume or no volume.",
          2                     text: "Lost coins only make everyone else's coins worth slightly more. Think of it as a donation to everyone.",

Explanation: print out all the lines with an empty character followed by
"text:" (to not get "context:" for example). Then sort them all.
Change double space to one (there is this error in one of our duplicity
quotes). Count the unique lines. Filter out the lines which were found
only once. The result is the two duplicates before this patch, exactly
as reported in #455.